* split nest.py into nest.py (API) and nesttool.py (command line tool). * modified setup.py to install all scripts and modules [not tested yet!]
